+++ src/biz/bokhorst/xprivacy/XPackageManager.java
@@ -31,7 +31,6 @@ public class XPackageManager extends XHook {
 
 	// public List<ApplicationInfo> getInstalledApplications(int flags)
 	// public List<PackageInfo> getInstalledPackages(int flags)
-	// public String[] getPackagesForUid(int uid)
 	// public List<PackageInfo> getPackagesHoldingPermissions(String[] permissions, int flags)
 	// public List<PackageInfo> getPreferredPackages(int flags)
 	// public List<ResolveInfo> queryBroadcastReceivers(Intent intent, int flags)
@@ -45,7 +44,7 @@ public class XPackageManager extends XHook {
 	// @formatter:on
 
 	private enum Methods {
-		getInstalledApplications, getInstalledPackages, getPackagesForUid, getPackagesHoldingPermissions, getPreferredPackages, queryBroadcastReceivers, queryContentProviders, queryIntentActivities, queryIntentActivityOptions, queryIntentContentProviders, queryIntentServices
+		getInstalledApplications, getInstalledPackages, getPackagesHoldingPermissions, getPreferredPackages, queryBroadcastReceivers, queryContentProviders, queryIntentActivities, queryIntentActivityOptions, queryIntentContentProviders, queryIntentServices
 	};
 
 	public static List<XHook> getInstances(Object instance) {
@@ -74,9 +73,6 @@ public class XPackageManager extends XHook {
 		if (mMethod == Methods.getInstalledApplications) {
 			if (param.getResult() != null && isRestricted(param))
 				param.setResult(filterApplicationInfo((List<ApplicationInfo>) param.getResult()));
-		} else if (mMethod == Methods.getPackagesForUid) {
-			if (param.getResult() != null && isRestricted(param))
-				param.setResult(new String[0]);
 		} else if (mMethod == Methods.getInstalledPackages || mMethod == Methods.getPackagesHoldingPermissions
 				|| mMethod == Methods.getPreferredPackages) {
 			if (param.getResult() != null && isRestricted(param))